GlobalOptimization

Documentation for GlobalOptimization.

GlobalOptimization.AbstractAlgorithmSpecificOptions

GlobalOptimization.AbstractCandidate

GlobalOptimization.AbstractEvaluator

GlobalOptimization.AbstractOptimizationProblem

GlobalOptimization.AbstractOptimizer

GlobalOptimization.AbstractOptions

GlobalOptimization.AbstractPopulation

GlobalOptimization.AsyncEvaluator

GlobalOptimization.BasicCandidate

GlobalOptimization.BasicEvaluator

GlobalOptimization.BatchEvaluator

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.FitnessAwareCandidate

GlobalOptimization.FixedDimensionSearchSpace

GlobalOptimization.GeneralOptions

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.PSO

GlobalOptimization.PSOCache

GlobalOptimization.PSOOptions

GlobalOptimization.RectangularSearchSpace

GlobalOptimization.Results

GlobalOptimization.SearchSpace

GlobalOptimization.SerialBatchEvaluator

GlobalOptimization.SingleEvaluator

GlobalOptimization.Swarm

GlobalOptimization.Swarm

GlobalOptimization.ThreadedBatchEvaluator

GlobalOptimization.SerialPSO

GlobalOptimization.Swarm_F32

GlobalOptimization.Swarm_F64

GlobalOptimization.ThreadedPSO

GlobalOptimization.candidates

GlobalOptimization.check_fitness!

GlobalOptimization.dimdelta

GlobalOptimization.dimmax

GlobalOptimization.dimmin

GlobalOptimization.dimrange

GlobalOptimization.display_status

GlobalOptimization.enforce_bounds!

GlobalOptimization.evaluate

GlobalOptimization.evaluate!

GlobalOptimization.evaluate!

GlobalOptimization.evaluate_fitness!

GlobalOptimization.exprate

GlobalOptimization.expscale

GlobalOptimization.feasible

GlobalOptimization.fitness

GlobalOptimization.get_display

GlobalOptimization.get_display_interval

GlobalOptimization.get_function_value_check

GlobalOptimization.get_general

GlobalOptimization.get_max_time

GlobalOptimization.initialize_fitness!

GlobalOptimization.initialize_uniform!

GlobalOptimization.intersection

GlobalOptimization.laplace

GlobalOptimization.laplace

GlobalOptimization.laplace

GlobalOptimization.numdims

GlobalOptimization.numdims

GlobalOptimization.optimize!

GlobalOptimization.search_space

GlobalOptimization.set_fitness!

GlobalOptimization.set_fitness!

GlobalOptimization.stalled

GlobalOptimization.step!

GlobalOptimization.update_global_best!

GlobalOptimization.update_inertia

GlobalOptimization.update_velocity!

GlobalOptimization.AbstractAlgorithmSpecificOptions — Type

AbstractAlgorithmSpecificOptions

Abstract type for algorithm specific options

GlobalOptimization.AbstractCandidate — Type

AbstractCandidate

Abstract type for a candidate

GlobalOptimization.AbstractEvaluator — Type

AbstractEvaluator

Abstract type for an evaluator. An evaluator is responsible for evaluating the fitness of a population or candidate.

GlobalOptimization.AbstractOptimizationProblem — Type

AbstractOptimizationProblem

Abstract type for optimization problems.

GlobalOptimization.AbstractOptimizer — Type

AbstractOptimizer

Abstract type of all optimization algorithms.

GlobalOptimization.AbstractOptions — Type

AbstractOptions

Abstract type for multiple options

GlobalOptimization.AbstractPopulation — Type

AbstractPopulation

Abstract type for a population of candidates.

GlobalOptimization.AsyncEvaluator — Type

AsyncEvaluator

Abstract type for an evaluator that evaluates the fitness of a single candidate asyncronously.

GlobalOptimization.BasicCandidate — Type

BasicCandidate

A basic candidate that knows who it is and its fitness

GlobalOptimization.BasicEvaluator — Type

BasicEvaluator

A basic evaluator that computes the fitness of a single candidate.

GlobalOptimization.BatchEvaluator — Type

BatchEvaluator

Abstract type for an evaluator that evaluates the fitness of an entire population.

GlobalOptimization.ContinuousRectangularSearchSpace — Type

ContinuousRectangularSearchSpace

A RectangularSearchSpace formed by a single continuous set.

GlobalOptimization.FitnessAwareCandidate — Type

FitnessAwareCandidate

Abstract type for a candidate that knows (or at least has the ability to know) its fitness.

Note: Name was selected to leave room for subtypes of AbstractCandidate that are not aware of their fitness (i.e., just a vector representing the candidate)

GlobalOptimization.FixedDimensionSearchSpace — Type

FixedDimensionSearchSpace

The base abstract type for a search space with a fixed finite number of dimensions. Applicable to the vast majority of optimization problems.

GlobalOptimization.GeneralOptions — Type

GeneralOptions{display, function_value_check}

General options for all optimizers

GlobalOptimization.OptimizationProblem — Type

OptimizationProblem

An optimization problem. Contains the objective function and search space.

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f, LB, UB)

Constructs an optimization problem with objective function f and a ContinuousRectangularSearchSpace defined by LB and UB.

GlobalOptimization.PSO — Type

PSO

Particle Swarm Optimization (PSO) algorithm.

GlobalOptimization.PSOCache — Type

PSOCache

Cache for PSO algorithm.

GlobalOptimization.PSOOptions — Type

PSOOptions <: AbstractAlgorithmSpecificOptions

Options for the PSO algorithm.

GlobalOptimization.RectangularSearchSpace — Type

RectangularSearchSpace

A FixedDimensionSearchSpace with N dimensional rectangle as the set of feasible points.

GlobalOptimization.Results — Type

Results

A simple struct for returning results.

GlobalOptimization.SearchSpace — Type

SearchSpace

The base abstract type for a Problem search space.

GlobalOptimization.SerialBatchEvaluator — Type

SerialBatchEvaluator

An evaluator that evaluates the fitness of a population in serial.

GlobalOptimization.SingleEvaluator — Type

SingleEvaluator

Abstract type for an evaluator that evaluates the fitness of a single candidate

GlobalOptimization.Swarm — Type

Swarm{T <: AbstractFloat} <: AbstractPopulation

A population of particles for the PSO algorithm.

GlobalOptimization.Swarm — Method

Swarm(num_particles::Integer, num_dims::Integer)

Constructs a Swarm with num_particles particles in num_dims dimensions.

GlobalOptimization.ThreadedBatchEvaluator — Type

ThreadedBatchEvaluator

An evaluator that evaluates the fitness of a population in parallel using multi-threading.

GlobalOptimization.SerialPSO — Method

SerialPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a serial PSO algorithm with the given options.

GlobalOptimization.Swarm_F32 — Method

Swarm_F32(num_particles::Integer, num_dims::Integer)

Constructs a Float32 Swarm with num_particles particles in num_dims dimensions.

GlobalOptimization.Swarm_F64 — Method

Swarm_F64(num_particles::Integer, num_dims::Integer)

Constructs a Float64 Swarm with num_particles particles in num_dims dimensions.

GlobalOptimization.ThreadedPSO — Method

ThreadedPSO

Constructs a threaded PSO algorithm with the given options.

GlobalOptimization.candidates — Method

candidates(pop::AbstractPopulation, [i::Integer])

Returns the candidates from a population. If i is specified, returns the i-th candidate.

GlobalOptimization.check_fitness! — Method

check_fitness!(pop::AbstractPopulation, options::Union{GeneralOptions,Val{true},Val{false}})

Checks the fitness of each candidate in the population pop to ensure that it is valid iff options <: Union{GeneralOptions{D,Val{true}}, Val{true}}, otherwise, does nothing.

GlobalOptimization.dimdelta — Method

dimdelta(ss::ContinuousRectangularSearchSpace, [i::Integer])

Returns the difference between the maximum and minimum values for the i-th dimension of ss. If i is not specified, returns a vector of all differences.

GlobalOptimization.dimmax — Method

dimmax(ss::ContinuousRectangularSearchSpace, [i::Integer])

Returns the maximum value for the i-th dimension of ss. If i is not specified, returns a vector of all maximum values.

GlobalOptimization.dimmin — Method

dimmin(ss::ContinuousRectangularSearchSpace, [i::Integer])

Returns the minimum value for the i-th dimension of ss. If i is not specified, returns a vector of all minimum values.

GlobalOptimization.dimrange — Method

dimrange(ss::ContinuousRectangularSearchSpace, [i::Integer])

Returns the range of values for the i-th dimension of ss. If i is not specified, returns a vector of all ranges.

GlobalOptimization.display_status — Method

display_status(time, iteration, stall_count, options)

Displays the status of the PSO algorithm.

GlobalOptimization.enforce_bounds! — Method

enforce_bounds!(swarm::Swarm{T}, evaluator::BatchEvaluator)

Enforces the bounds of the search space on each candidate in the swarm swarm. If a candidate

GlobalOptimization.evaluate! — Method

evaluate!(pop::AbstractPopulation, evaluator::BatchEvaluator)

Evaluates the fitness of a population using the given evaluator.

GlobalOptimization.evaluate! — Method

evaluate!(can::FitnessAwareCandidate, evaluator::BasicEvaluator)

Evaluates the fitness of a candidat using the given evaluator

GlobalOptimization.evaluate — Method

evaluate(prob::OptimizationProblem, x::AbstractArray)

Evaluates the objective function f of the optimization problem prob at x.

GlobalOptimization.evaluate_fitness! — Method

evaluate_fitness!(swarm::Swarm{T}, evaluator::BatchEvaluator{T})

Evaluates the fitness of each candidate in the swarm swarm using the evaluator. Updates the swarms best candidates if any are found.

GlobalOptimization.exprate — Method

exprate(λ::AbstractFloat)

Generate an exponential random variable with rate λ.

GlobalOptimization.expscale — Method

expscale(β::AbstractFloat)

Generate an exponential random variable with scale β.

GlobalOptimization.feasible — Method

feasible(x, ss::ContinuousRectangularSearchSpace)

Returns true if the point x is feasible in the search space ss, otherwise returns false.

GlobalOptimization.fitness — Method

fitness(pop::AbstractPopulation, [i::Integer])

Returns the fitness of the candidates from a population. If i is specified, returns the i-th fitness.

GlobalOptimization.get_display — Method

get_display(opts::AbstractOptions)

Returns the display option from an options type.

GlobalOptimization.get_display_interval — Method

get_display_interval(opts::AbstractAlgorithmSpecificOptions)

Returns the display interval from an algorithm options type.

GlobalOptimization.get_function_value_check — Method

get_function_value_check(opts::AbstractAlgorithmSpecificOptions)

Returns the function value check option from an algorithm options type.

GlobalOptimization.get_general — Method

get_general(opts::AbstractAlgorithmSpecificOptions)

Returns the general options from an algorithm options type.

GlobalOptimization.get_max_time — Method

get_max_time(opts::AbstractAlgorithmSpecificOptions)

Returns the max time option from an algorithm options type.

GlobalOptimization.initialize_fitness! — Method

initialize_fitness!(swarm::Swarm{T}, evaluator::BatchEvaluator{T})

Initializes the fitness of each candidate in the swarm swarm using the evaluator.

GlobalOptimization.initialize_uniform! — Method

initialize_uniform!(swarm::Swarm{T}, search_space::ContinuousRectangularSearchSpace{T})

Initializes the swarm swarm with a uniform particle distribution in the search space.

GlobalOptimization.intersection — Method

intersection(ss1, ss2)

Returns the intersection of the two search spaces ss1 and ss2 as a new search sapce

GlobalOptimization.laplace — Method

laplace(b)

Generate a Laplace random variable with location 0.0 and scale b.

GlobalOptimization.laplace — Method

laplace([::Type{T}]) where T <: AbstractFloat

Generate a Laplace random variable with location 0.0 and scale 1.0. If T is specified, the random variable will be of type T.

GlobalOptimization.laplace — Method

laplace(μ::AbstractFloat, b::Real)

Generate a Laplace random variable with location μ and scale b.

GlobalOptimization.numdims — Method

numdims(ss::ContinuousRectangularSearchSpace)

Returns the number of dimensions in the search space ss.

GlobalOptimization.numdims — Method

numdims(prob::OptimizationProblem)

Returns the number of dimensions of the decision vector of the optimization problem prob.

GlobalOptimization.optimize! — Method

optimize!(opt::AbstractOptimizer)

Perform optimization using the optimizer opt. Returns the results of the optimization.

GlobalOptimization.search_space — Method

search_space(prob::OptimizationProblem)

Returns the search space of the optimization problem prob.

GlobalOptimization.set_fitness! — Method

set_fitness(pop::AbstractPopulation, fitness, [i::Integer])

Sets the fitness of the candidates from a population. If i is specified, sets the i-th fitness.

GlobalOptimization.set_fitness! — Method

set_fitness!(candidate,fitness)

Sets the fitness of a candidate.

GlobalOptimization.stalled — Method

stalled(pso::PSO, stall_value)

Returns true if the PSO algorithm pso is stalled, false otherwise.

GlobalOptimization.step! — Method

step!(swarm::Swarm)

Steps the swarm swarm forward one iteration.

GlobalOptimization.update_global_best! — Method

update_global_best!(pso::PSO)

Updates the global best candidate in the PSO algorithm pso if a better candidate is found.

GlobalOptimization.update_inertia — Method

update_inertia(inertia, range, stall_count)

Returns new inertia weight based on the current 'inertia', the range, and stall_count.

GlobalOptimization.update_velocity! — Method

update_velocity!(swarm::Swarm{T}, cache::Cache, ns::Integer, w, y1, y2)

Updates the velocity of each candidate in the swarm swarm,