Public API Documentation

Documentation for GlobalOptimization's public interface.

Contents

Contents

Index

Optimization Problem

Search Space

Optimization

Algorithms

Particle Swarm Optimization

Differential Evolution

Index

GlobalOptimization.Best1

GlobalOptimization.Best2

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.CurrentToBest1

GlobalOptimization.CurrentToBest2

GlobalOptimization.CurrentToRand1

GlobalOptimization.CurrentToRand2

GlobalOptimization.MutationParameters

GlobalOptimization.MutationParameters

GlobalOptimization.MutationParameters

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.RadiusLimitedSelector

GlobalOptimization.Rand1

GlobalOptimization.Rand2

GlobalOptimization.RandToBest1

GlobalOptimization.RandToBest2

GlobalOptimization.SelfMutationParameters

GlobalOptimization.SelfMutationParameters

GlobalOptimization.SimpleSelector

GlobalOptimization.Unified

GlobalOptimization.PolyesterDE

GlobalOptimization.PolyesterPSO

GlobalOptimization.SerialDE

GlobalOptimization.SerialPSO

GlobalOptimization.ThreadedDE

GlobalOptimization.ThreadedPSO

GlobalOptimization.optimize!

Optimization Problem

GlobalOptimization.OptimizationProblem — Type

OptimizationProblem{has_penalty, SS, F, G}

An optimization problem. Contains the objective function and search space.

Fields

f::F: The objective function.

g!::G: The gradient of the objective function.

ss::SS: The search space.

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f, [g], LB, UB)

Constructs an optimization problem with objective function f, optional gradient g, and a ContinuousRectangularSearchSpace defined by LB and UB.

Arguments

f::F: The objective function.

LB::AbstractVector{<:Real}: The lower bounds of the search space.

UB::AbstractVector{<:Real}: The upper bounds of the search space.

Returns

OptimizationProblem{ContinuousRectangularSearchSpace, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> prob = OptimizationProblem(f, LB, UB)
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem{has_penalty}(f::F, [g::G], ss::SS)

Constructs an optimization problem with objective function f, optional gradient g, and search space ss. If has_penalty is specified as true, then the objective function must return a Tuple{T,T} for a given x of type AbstractArray{T}.

Arguments

f::F: The objective function.

g::G: The gradient of the objective function.

ss::SS: The search space.

Returns

OptimizationProblem{SS, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = OptimizationProblem(f, ss)
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f, [g], ss)

Constructs an optimization problem with objective function f, optimal gradient g, and a search space.

Arguments

f::F: The objective function.

g::G: The gradient of the objective function.

ss::SS: The search space.

Returns

OptimizationProblem{ContinuousRectangularSearchSpace, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> prob = OptimizationProblem(f, ContinuousRectangularSearchSpace(LB, UB))
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

Search Space

GlobalOptimization.ContinuousRectangularSearchSpace — Type

ContinuousRectangularSearchSpace{T <: AbstractFloat}

A RectangularSearchSpace formed by a single continuous set.

Fields

dimmin::Vector{T}: A vector of minimum values for each dimension.

dimmax::Vector{T}: A vector of maximum values for each dimension.

dimdelta::Vector{T}: A vector of the difference between the maximum and minimum values for each dimension.

GlobalOptimization.ContinuousRectangularSearchSpace — Method

ContinuousRectangularSearchSpace(dimmin::AbstractVector{T}, dimmax::AbstractVector{T})

Constructs a new ContinuousRectangularSearchSpace with minimum values dimmin and maximum values dimmax.

Arguments

dimmin::AbstractVector{T}: A vector of minimum values for each dimension.

dimmax::AbstractVector{T}: A vector of maximum values for each dimension.

Returns

ContinuousRectangularSearchSpace{T}

Examples

julia> using GlobalOptimization;
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB)
ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0])

Optimization

GlobalOptimization.optimize! — Method

optimize!(opt::AbstractOptimizer)

Perform optimization using the optimizer opt. Returns the results of the optimization.

Arguments

opt::AbstractOptimizer: The optimizer to use.

Returns

Results: The results of the optimization.

Algorithms

Particle Swarm Optimization

GlobalOptimization.PolyesterPSO — Method

PolyesterPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a PolyesterBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractOptimizationProblem{SS}: The optimization problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

GlobalOptimization.SerialPSO — Method

SerialPSO(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a SerialBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

GlobalOptimization.ThreadedPSO — Method

ThreadedPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a ThreadedBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractOptimizationProblem{SS}: The optimization problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

Differential Evolution

GlobalOptimization.PolyesterDE — Method

PolyesterDE(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Construct a serial Differential Evolution (DE) algorithm with the given options that will employ a PolyesterBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_candidates::Integer=100: The number of candidates in the population.

mutation_params::MP=SelfMutationParameters(Rand1()): The mutation strategy parameters.

crossover_params::CP=BinomialCrossoverParameters(0.6): The crossover strategy parameters.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace}=nothing: The initial bounds for the search space.

max_iterations::Integer=1000: The maximum number of iterations.

max_time::Real=60.0: The maximum time to run the algorithm.

function_tolerance::Real=1e-6: The function tolerance for the stall condition.

max_stall_time::Real=60.0: The maximum stall time for the stall condition.

max_stall_iterations::Integer=100: The maximum number of stall iterations for the stall condition.

min_cost::Real=-Inf: The minimum cost for the algorithm to stop.

function_value_check::Bool=true: Whether to check the function value.

display::Bool=true: Whether to display the algorithm status.

display_interval::Integer=1: The interval at which to display the algorithm status.

GlobalOptimization.SerialDE — Method

SerialDE(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Construct a serial Differential Evolution (DE) algorithm with the given options that will employ a SerialBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_candidates::Integer=100: The number of candidates in the population.

mutation_params::MP=SelfMutationParameters(Rand1()): The mutation strategy parameters.

crossover_params::CP=BinomialCrossoverParameters(0.6): The crossover strategy parameters.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace}=nothing: The initial bounds for the search space.

max_iterations::Integer=1000: The maximum number of iterations.

max_time::Real=60.0: The maximum time to run the algorithm.

function_tolerance::Real=1e-6: The function tolerance for the stall condition.

max_stall_time::Real=60.0: The maximum stall time for the stall condition.

max_stall_iterations::Integer=100: The maximum number of stall iterations for the stall condition.

min_cost::Real=-Inf: The minimum cost for the algorithm to stop.

function_value_check::Bool=true: Whether to check the function value.

display::Bool=true: Whether to display the algorithm status.

display_interval::Integer=1: The interval at which to display the algorithm status.

GlobalOptimization.ThreadedDE — Method

ThreadedDE(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Construct a serial Differential Evolution (DE) algorithm with the given options that will employ a ThreadedBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_candidates::Integer=100: The number of candidates in the population.

mutation_params::MP=SelfMutationParameters(Rand1()): The mutation strategy parameters.

crossover_params::CP=BinomialCrossoverParameters(0.6): The crossover strategy parameters.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace}=nothing: The initial bounds for the search space.

max_iterations::Integer=1000: The maximum number of iterations.

max_time::Real=60.0: The maximum time to run the algorithm.

function_tolerance::Real=1e-6: The function tolerance for the stall condition.

max_stall_time::Real=60.0: The maximum stall time for the stall condition.

max_stall_iterations::Integer=100: The maximum number of stall iterations for the stall condition.

min_cost::Real=-Inf: The minimum cost for the algorithm to stop.

function_value_check::Bool=true: Whether to check the function value.

display::Bool=true: Whether to display the algorithm status.

display_interval::Integer=1: The interval at which to display the algorithm status.

Mutation Parameters

GlobalOptimization.MutationParameters — Type

MutationParameters{
    AS<:AbstractAdaptationStrategy,
    MS<:AbstractMutationStrategy,
    S<:AbstractSelector,
    D,
}

The parameters for a DE mutation strategy that applies to all current and future candidates in the population.

Fields

F1::Float64: The F₁ weight in the unified mutation strategy.

F2::Float64: The F₂ weight in the unified mutation strategy.

F3::Float64: The F₃ weight in the unified mutation strategy.

F4::Float64: The F₄ weight in the unified mutation strategy.

sel<:AbstractSelector: The selector used to select the candidates considered in mutation.

dist<:Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value.

GlobalOptimization.MutationParameters — Method

MutationParameters(F1, F2, F3, F4; sel=SimpleSelector())

Creates a MutationParameters object with the specified (constant) mutation parameters. These constant mutation parameters are used for all candidates in the population and define a unified mutation strategy as defined in Ji Qiang and Chad Mitchell "A Unified Differential Evolution Algorithm for Global Optimization," 2014, https://www.osti.gov/servlets/purl/1163659

Arguments

F1::Float64: The F₁ weight in the unified mutation strategy.

F2::Float64: The F₂ weight in the unified mutation strategy.

F3::Float64: The F₃ weight in the unified mutation strategy.

F4::Float64: The F₄ weight in the unified mutation strategy.

Keyword Arguments

sel::AbstractSelector: The selector used to select the candidates considered in mutation. Defaults to SimpleSelector().

GlobalOptimization.MutationParameters — Method

MutationParameters(
    strategy::MS;
    dist=default_mutation_dist,
    sel=SimpleSelector(),
)

Creates a MutationParameters object with the specified mutation strategy with mutation parameter random adaptation. The mutation parameters are adaptively sampled from the provided dist, clamped to the range (0, 1].

Arguments

strategy::MS: The mutation strategy to use. This should be one of the mutation strategies defined in this module (e.g., Rand1, Best2, etc.).

Keyword Arguments

dist::Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters each iteration. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value. Defaults to GlobalOptimization.default_mutation_dist, which is a mixture model comprised of two Cauchy distributions, with probability density given by:

$f_{mix}(x; \mu, \sigma) = 0.5 f(x;\mu_1,\sigma_1) + 0.5 f(x;\mu_2,\sigma_2)$.

where $\mu = \{0.65, 1.0\}$ and $\sigma = \{0.1, 0.1\}$.

sel::AbstractSelector: The selector used to select the candidates considered in mutation. Defaults to SimpleSelector().

GlobalOptimization.SelfMutationParameters — Type

SelfMutationParameters{
    AS<:AbstractAdaptationStrategy,
    MS<:AbstractMutationStrategy,
    S<:AbstractSelector,
    D,
}

The parameters for a DE mutation strategy that applies a mutation strategy with unique parameters for each candidate in the population.

Fields

Fs::Vector{SVector{4,Float64}}: The mutation parameters for each candidate in the population. Each element of the vector is an SVector{4} containing the F₁, F₂, F₃, and F₄ weights for the unified mutation strategy.

sel<:AbstractSelector: The selector used to select the candidates considered in mutation.

dist<:Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value.

GlobalOptimization.SelfMutationParameters — Method

SelfMutationParameters(
    strategy::MS;
    dist=default_mutation_dist,
    sel=SimpleSelector(),
)

Creates a SelfMutationParameters object with the specified mutation strategy and mutation parameter random adaptation. The mutation parameters are adaptively sampled from the provided dist, clamped to the range (0, 1].

Arguments

strategy::MS: The mutation strategy to use. This should be one of the mutation strategies defined in this module (e.g., Rand1, Best2, etc.).

Keyword Arguments

dist::Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters each iteration. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value. Defaults to GlobalOptimization.default_mutation_dist, which is a mixture model comprised of two Cauchy distributions, with probability density given by:

$f_{mix}(x; \mu, \sigma) = 0.5 f(x;\mu_1,\sigma_1) + 0.5 f(x;\mu_2,\sigma_2)$.

where $\mu = \{0.65, 1.0\}$ and $\sigma = \{0.1, 0.1\}$.

sel::AbstractSelector: The selector used to select the candidates considered in mutation. Defaults to SimpleSelector().

GlobalOptimization.Rand1 — Type

Rand1

The DE/rand/1 mutation strategy.

GlobalOptimization.Rand2 — Type

Rand2

The DE/rand/2 mutation strategy.

GlobalOptimization.Best1 — Type

Best1

The DE/best/1 mutation strategy.

GlobalOptimization.Best2 — Type

Best2

The DE/best/2 mutation strategy.

GlobalOptimization.CurrentToBest1 — Type

CurrentToBest1

The DE/current-to-best/1 mutation strategy.

GlobalOptimization.CurrentToBest2 — Type

CurrentToBest2

The DE/current-to-best/2 mutation strategy.

GlobalOptimization.CurrentToRand1 — Type

CurrentToRand1

The DE/current-to-rand/1 mutation strategy.

GlobalOptimization.CurrentToRand2 — Type

CurrentToRand2

The DE/current-to-rand/2 mutation strategy.

GlobalOptimization.RandToBest1 — Type

RandToBest1

The DE/rand-to-best/1 mutation strategy.

GlobalOptimization.RandToBest2 — Type

RandToBest2

The DE/rand-to-best/2 mutation strategy.

GlobalOptimization.Unified — Type

Unified

The unified DE mutation strategy proposed by Ji Qiang and Chad Mitchell in "A Unified Differential Evolution Algorithm for Global Optimization," 2014, https://www.osti.gov/servlets/purl/1163659

GlobalOptimization.SimpleSelector — Type

SimpleSelector

A selector that simply selects all candidates in the population.

GlobalOptimization.RadiusLimitedSelector — Type

RadiusLimitedSelector

A selector that selects candidates within a given radius of the target candidate.

For example, for population size of 10 and a radius of 2, the following will be selected for the given target indices:

target = 5 will select [3, 4, 5, 6, 7]

target = 1 will select [9, 10, 1, 2, 3]

target = 9 will select [7, 8, 9, 10, 1]

DE Crossover Strategies