Internals

Listing all non-exported types and functions here for now, but split off categories onto separate pages in the future!

GlobalOptimization.AbstractAlgorithmSpecificOptions — Type

AbstractAlgorithmSpecificOptions

Abstract type for algorithm specific options

GlobalOptimization.AbstractCandidate — Type

AbstractCandidate

Abstract type for a candidate

GlobalOptimization.AbstractEvaluator — Type

AbstractEvaluator

Abstract type for an evaluator. An evaluator is responsible for evaluating the fitness of a population or candidate.

GlobalOptimization.AbstractMBHDistribution — Type

AbstractMBHDistribution{T}

Abstract type for MBH distributions.

GlobalOptimization.AbstractNonlinearEquationProblem — Type

AbstractNonlinearEquationProblem

Abstract type for problems involving a set of nonlinear equations to solve.

GlobalOptimization.AbstractOptimizationProblem — Type

AbstractOptimizationProblem

Abstract type for optimization problems.

GlobalOptimization.AbstractOptimizer — Type

AbstractOptimizer

Abstract type of all optimization algorithms.

GlobalOptimization.AbstractOptions — Type

AbstractOptions

Abstract type for multiple options

GlobalOptimization.AbstractPopulation — Type

AbstractPopulation

Abstract type for a population of candidates.

GlobalOptimization.AbstractProblem — Type

AbstractProblem

Abstract type for all solvable problems.

GlobalOptimization.AsyncEvaluator — Type

AsyncEvaluator

Abstract type for an evaluator that evaluates the fitness of a single candidate asyncronously.

GlobalOptimization.BatchEvaluator — Type

BatchEvaluator

Abstract type for an evaluator that evaluates the fitness of an entire population.

GlobalOptimization.DE — Type

DE

Differential Evolution (DE) algorithm.

GlobalOptimization.DEBasePopulation — Type

DEBasePopulation{T <: AbstractFloat} <: AbstractPopulation{T}

The base representation of some population for the DE algorithms.

GlobalOptimization.DECache — Type

DECache

Cache for the DE algorithm.

GlobalOptimization.DEOptions — Type

DEOptions

Options for the Differential Evolution (DE) algorithms.

Fields:

general<:GeneralOptions: The general options.

pop_init_method<:AbstractPopulationInitialization: The population initialization method.

mutation_params<:AbstractMutationParameters: The mutation strategy parameters.

crossover_params<:AbstractCrossoverParameters: The crossover strategy parameters.

initial_space<:Union{Nothing,ContinuousRectangularSearchSpace}: The initial space to initialize the population.

max_iterations::Int: The maximum number of iterations.

function_tolerance::Float64: The function tolerance for the stall condition.

max_stall_time::Float64: The maximum stall time for the stall condition.

max_stall_iterations::Int: The maximum number of stall iterations for the stall condition.

GlobalOptimization.DEPopulation — Type

DEPopulation{T <: AbstractFloat} <: AbstractPopulation{T}

The full population representation for the DE algorithms, including both the candidates and the mutants.

GlobalOptimization.DEPopulation — Method

DEPopulation(num_candidates::Integer, num_dims::Integer)

Constructs a DEPopulation with num_candidates candidates in num_dims dimensions.

GlobalOptimization.FeasibilityHandlingEvaluator — Type

FeasibilityHandlingEvaluator

An evaluator that handled a functions returned infeasibility penalty

GlobalOptimization.FixedDimensionSearchSpace — Type

FixedDimensionSearchSpace

The base abstract type for a search space with a fixed finite number of dimensions. Applicable to the vast majority of optimization problems.

GlobalOptimization.GeneralOptions — Type

GeneralOptions{display, function_value_check}

General options for all optimizers

GlobalOptimization.MBH — Type

MBH

Monotonic Basin Hopping (MBH) algorithm.

This implementation employs a single candidate rather than a population.

GlobalOptimization.MBH — Method

MBH(prob::AbstractOptimizationProblem{SS})

GlobalOptimization.MBHAdaptiveDistribution — Type

MBHAdaptiveDistribution{T}

Adaptive distribution for MBH.

GlobalOptimization.MBHOptions — Type

MBHOptions <: AbstractAlgorithmSpecificOptions

Options for the Monotonic Basin Hopping (MBH) algorithm.

GlobalOptimization.MBHStaticDistribution — Type

MBHStaticDistribution{T}

Static distribution for MBH.

GlobalOptimization.MBHStepMemory — Type

MBHStepMemory{T}

Memory about MBH accepted steps.

GlobalOptimization.NonlinearLeastSquaresProblem — Type

NonlinearLeastSquaresProblem{has_penalty, SS, F, G}

A nonlinear least squares problem. Contains the nonlinear equations and search space.

Fields

f::F: The nonlinear equations.

g!::G: The jacobian of the nonlinear equations.

ss::SS: The search space.

n::Int: The number of residuals.

GlobalOptimization.NonlinearLeastSquaresProblem — Method

NonlinearLeastSquaresProblem(f, [g], LB, UB)

Constructs a nonlinear least squares problem with nonlinear function f, optional Jacobian g, and a search space.

Arguments

f::F: The nonlinear function.

g::G: The Jacobian of the nonlinear function.

LB::AbstractVector{<:Real}: The lower bounds of the search space.

UB::AbstractVector{<:Real}: The upper bounds of the search space.

Returns

NonlinearLeastSquaresProblem{has_penalty, ContinuousRectangularSearchSpace, F, G}

Examples

julia> using GlobalOptimization;
julia> f(x) = [x[1] - x[3], x[2] - x[3]]
julia> LB = [-5.0, -5.0, -5.0];
julia> UB = [ 5.0, 5.0, 5.0];
julia> prob = NonlinearLeastSquaresProblem(f, ss, LB, UB, 2)
NonlinearLeastSquaresProblem{Val{false}(), ContinuousRectangularSearchSpace{Float64}, typeof(f), Nothing}(f, nothing, ContinuousRectangularSearchSpace{Float64}([-5.0, -5.0, -5.0], [5.0, 5.0, 5.0], [10.0, 10.0, 10.0]), 2)

GlobalOptimization.NonlinearLeastSquaresProblem — Method

NonlinearLeastSquaresProblem{has_penalty}(f::F, [g::G], ss::SS, num_resid::Int)

Constructs a nonlinear least squares problem with nonlinear functions f, optional jacobian g, and search space ss. If has_penalty is specified as true, then the nonlinear function must return a Tuple{AbstractArray{T},T} for a given x of type AbstractArray{T}.

Arguments

f::F: The nonlinear function.

g::G: The Jacobian of the nonlinear function.

ss::SS: The search space.

num_resid::Int: The number of residuals.

Returns

NonlinearLeastSquaresProblem{has_penalty, SS, F, G}

Examples

julia> using GlobalOptimization;
julia> f(x) = [x[1] - x[3], x[2] - x[3]]
julia> LB = [-5.0, -5.0, -5.0];
julia> UB = [ 5.0, 5.0, 5.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = NonlinearLeastSquaresProblem(f, ss, 2)
NonlinearLeastSquaresProblem{Val{false}(), ContinuousRectangularSearchSpace{Float64}, typeof(f), Nothing}(f, nothing, ContinuousRectangularSearchSpace{Float64}([-5.0, -5.0, -5.0], [5.0, 5.0, 5.0], [10.0, 10.0, 10.0]), 2)

GlobalOptimization.NonlinearLeastSquaresProblem — Method

NonlinearLeastSquaresProblem(f, [g], ss)

Constructs a nonlinear least squares problem with nonlinear function f, optional Jacobian g, and a search space.

Arguments

f::F: The nonlinear function.

g::G: The Jacobian of the nonlinear function.

ss::SS: The search space.

Returns

NonlinearLeastSquaresProblem{has_penalty, ContinuousRectangularSearchSpace, F, G}

Examples

julia> using GlobalOptimization;
julia> f(x) = [x[1] - x[3], x[2] - x[3]]
julia> LB = [-5.0, -5.0, -5.0];
julia> UB = [ 5.0, 5.0, 5.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = NonlinearLeastSquaresProblem(f, ss, 2)
NonlinearLeastSquaresProblem{Val{false}(), ContinuousRectangularSearchSpace{Float64}, typeof(f), Nothing}(f, nothing, ContinuousRectangularSearchSpace{Float64}([-5.0, -5.0, -5.0], [5.0, 5.0, 5.0], [10.0, 10.0, 10.0]), 2)

GlobalOptimization.NonlinearProblem — Type

NonlinearProblem{has_penalty, SS, F, G}

A nonlinear problem. Contains the nonlinear equations and search space.

Fields

f::F: The nonlinear equations.

g!::G: The jacobian of the nonlinear equations.

ss::SS: The search space.

GlobalOptimization.NonlinearProblem — Method

NonlinearProblem(f, [g], LB, UB)

Constructs a nonlinear problem with nonlinear function f, optional Jacobian g, and a continuous rectangular search space defined by the bounds LB and UB.

Arguments

f::F: The nonlinear function.

g::G: The Jacobian of the nonlinear function.

LB::AbstractVector{<:Real}: The lower bounds of the search space.

UB::AbstractVector{<:Real}: The upper bounds of the search space.

Returns

NonlinearProblem{has_penalty, ContinuousRectangularSearchSpace, F, G}

Examples

julia> using GlobalOptimization;
julia> f(x) = [x[1] - 2.0, x[2] - 2.0]
julia> LB = [-5.0, -5.0];
julia> UB = [ 5.0, 5.0];
julia> prob = NonlinearProblem(f, LB, UB)
NonlinearProblem{Val{false}(), ContinuousRectangularSearchSpace{Float64}, typeof(f), Nothing}(f, nothing, ContinuousRectangularSearchSpace{Float64}([-5.0, -5.0], [5.0, 5.0], [10.0, 10.0]))

GlobalOptimization.NonlinearProblem — Method

NonlinearProblem{has_penalty}(f::F, [g::G], ss::SS)

Constructs a nonlinear problem with nonlinear functions f, optional jacobian g, and search space ss. If has_penalty is specified as true, then the nonlinear function must return a Tuple{AbstractArray{T},T} for a given x of type AbstractArray{T}.

Arguments

f::F: The nonlinear function.

g::G: The Jacobian of the nonlinear function.

ss::SS: The search space.

Returns

NonlinearProblem{has_penalty, SS, F, G}

Examples

julia> using GlobalOptimization;
julia> f(x) = [x[1] - 2.0, x[2] - 2.0]
julia> LB = [-5.0, -5.0];
julia> UB = [ 5.0, 5.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = NonlinearProblem(f, ss)
NonlinearProblem{Val{false}(), ContinuousRectangularSearchSpace{Float64}, typeof(f), Nothing}(f, nothing, ContinuousRectangularSearchSpace{Float64}([-5.0, -5.0], [5.0, 5.0], [10.0, 10.0]))

GlobalOptimization.NonlinearProblem — Method

NonlinearProblem(f, [g], ss)

Constructs a nonlinear problem with nonlinear function f, optional Jacobian g, and a search space.

Arguments

f::F: The nonlinear function.

g::G: The Jacobian of the nonlinear function.

ss::SS: The search space.

Returns

NonlinearProblem{has_penalty, ContinuousRectangularSearchSpace, F, G}

Examples

julia> using GlobalOptimization;
julia> f(x) = [x[1] - 2.0, x[2] - 2.0]
julia> LB = [-5.0, -5.0];
julia> UB = [ 5.0, 5.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = NonlinearProblem(f, ss)
NonlinearProblem{Val{false}(), ContinuousRectangularSearchSpace{Float64}, typeof(f), Nothing}(f, nothing, ContinuousRectangularSearchSpace{Float64}([-5.0, -5.0], [5.0, 5.0], [10.0, 10.0]))

GlobalOptimization.PSO — Type

PSO

Particle Swarm Optimization (PSO) algorithm.

GlobalOptimization.PSOCache — Type

PSOCache

Cache for PSO algorithm.

GlobalOptimization.PSOOptions — Type

PSOOptions <: AbstractAlgorithmSpecificOptions

Options for the PSO algorithm.

GlobalOptimization.PolyesterBatchEvaluator — Type

PolyesterBatchEvaluator

An evaluator that evaluates the fitness of a population in parallel using multi-threading using Polyester.jl.

GlobalOptimization.RectangularSearchSpace — Type

RectangularSearchSpace

A FixedDimensionSearchSpace with N dimensional rectangle as the set of feasible points.

GlobalOptimization.Results — Type

Results{T}

A simple struct for returning results.

Fields

fbest::T: The best function value found.

xbest::Vector{T}: The best candidate found.

iters::Int: The number of iterations performed.

time::Float64: The time taken to perform the optimization in seconds.

exitFlag::Int: The exit flag of the optimization.

GlobalOptimization.Results — Method

Results(fbest::T, xbest::AbstractVector{T}, iters, time, exitFlag)

Constructs a new Results struct.

Arguments

fbest::T: The best function value found.

xbest::AbstractVector{T}: The best candidate found.

iters::Int: The number of iterations performed.

time::AbstractFloat: The time taken to perform the optimization in seconds.

exitFlag::Int: The exit flag of the optimization.

Returns

Results{T}

GlobalOptimization.SearchSpace — Type

SearchSpace

The base abstract type for a Problem search space.

GlobalOptimization.SerialBatchEvaluator — Type

SerialBatchEvaluator

An evaluator that evaluates the fitness of a population in serial.

GlobalOptimization.SingleEvaluator — Type

SingleEvaluator

Abstract type for an evaluator that evaluates the fitness of a single candidate

GlobalOptimization.Swarm — Type

Swarm{T <: AbstractFloat} <: AbstractPopulation

A population of particles for the PSO algorithm.

GlobalOptimization.Swarm — Method

Swarm(num_particles::Integer, num_dims::Integer)

Constructs a Swarm with num_particles particles in num_dims dimensions.

GlobalOptimization.ThreadedBatchEvaluator — Type

ThreadedBatchEvaluator

An evaluator that evaluates the fitness of a population in parallel using multi-threading.

GlobalOptimization.UniformInitialization — Type

UniformInitialization

Initializes a population from a uniform distribution in the search space.

Base.eachindex — Method

eachindex(pop::AbstractPopulation)

Returns an iterator for the indices of the population.

Base.length — Method

length(pop::AbstractPopulation)

Returns the number of candidates in the population.

Base.size — Method

size(pop::AbstractPopulation)

Returns the size of the population.

GlobalOptimization.DEPopulation_F64 — Method

DEPopulation_F64(num_candidates::Integer, num_dims::Integer)

Constructs a Float64 DEPopulation with num_candidate candidates in num_dims dimensions.

GlobalOptimization.Swarm_F32 — Method

Swarm_F32(num_particles::Integer, num_dims::Integer)

Constructs a Float32 Swarm with num_particles particles in num_dims dimensions.

GlobalOptimization.Swarm_F64 — Method

Swarm_F64(num_particles::Integer, num_dims::Integer)

Constructs a Float64 Swarm with num_particles particles in num_dims dimensions.

GlobalOptimization.candidate — Method

candidate(c::AbstractCandidate)

Returns the candidate c.

GlobalOptimization.candidates — Method

candidates(pop::AbstractPopulation, [i::Integer])

Returns the candidates from a population. If i is specified, returns the i-th candidate.

GlobalOptimization.check_fitness! — Method

check_fitness!(c::AbstractCandidate, options::Union{GeneralOptions,Val{true},Val{false}})

Checks the fitness of the candidate c to ensure that it is valid iff options <: Union{GeneralOptions{D,Val{true}}, Val{true}}, otherwise, does nothing.

GlobalOptimization.check_fitness! — Method

check_fitness!(pop::AbstractPopulation, options::Union{GeneralOptions,Val{true},Val{false}})

Checks the fitness of each candidate in the population pop to ensure that it is valid iff options <: Union{GeneralOptions{D,Val{true}}, Val{true}}, otherwise, does nothing.

GlobalOptimization.crossover! — Method

crossover!(population::DEPopulation{T}, crossover_params, search_space)

Performs the crossover operation on the population population using the DE crossover strategy.

This function also ensures that, after crossover, the mutants are within the search space.

GlobalOptimization.dim_delta — Method

dim_delta(ss::ContinuousRectangularSearchSpace{T}, [i::Integer])

Returns the difference between the maximum and minimum values for the i-th dimension of ss. If i is not specified, returns a vector of all differences.

Arguments

ss::ContinuousRectangularSearchSpace{T}`

i::Integer: the dimension to return the difference between the maximum and minimum values for.

Returns

T or Vector{T} the difference between the maximum and minimum values for the i-th dimension of ss or a vector of all differences if i not provided.

GlobalOptimization.dim_max — Method

dim_max(ss::ContinuousRectangularSearchSpace{T}, [i::Integer])

Returns the maximum value for the i-th dimension of ss. If i is not specified, returns a vector of all maximum values.

Arguments

ss::ContinuousRectangularSearchSpace{T}

i::Integer: the dimension to return the maximum value for.

Returns

T or Vector{T}: the minimum value for the i-th dimension of ss or a vector of all minimum values if i not provided.

GlobalOptimization.dim_min — Method

dim_min(ss::ContinuousRectangularSearchSpace{T}, [i::Integer])

Returns the minimum value for the i-th dimension of ss. If i is not specified, returns a vector of all minimum values.

Arguments

ss::RontinuousRectangularSearchSpace{T}

i::Integer: the dimension to return the minimum value for.

Returns

T or Vector{T}: the minimum value for the i-th dimension of ss or a vector of all minimum values if i not provided.

GlobalOptimization.dim_range — Method

dim_range(ss::ContinuousRectangularSearchSpace{T}, [i::Integer])

Returns the range of values for the i-th dimension of ss. If i is not specified, returns a vector of all ranges.

Arguments

ss::ContinuousRectangularSearchSpace{T}

i::Integer: the dimension to return the range of values for.

Returns

Tuple{T, T} or Vector{Tuple{T, T}}: the range of values for the i-th dimension of ss or a vector of all ranges if i not provided.

GlobalOptimization.display_status — Method

display_status(time, iteration, stall_count, options)

Displays the status of the PSO algorithm.

GlobalOptimization.draw_step! — Method

draw_step!(step::AbstractVector{T}, dist::AbstractMBHDistribution{T})

Draws a step from the distribution dist and stores it in step.

GlobalOptimization.draw_update! — Method

draw_update!(hopper::BasicHopper{T}, distribution::AbstractMBHDistribution{T})

Draws a perterbation from distribution and updates candidate for the hopper hopper.

GlobalOptimization.enforce_bounds! — Method

enforce_bounds!(swarm::Swarm{T}, evaluator::BatchEvaluator)

Enforces the bounds of the search space on each candidate in the swarm swarm. If a candidate

GlobalOptimization.evaluate! — Method

evaluate!(pop::AbstractPopulation, evaluator::BatchEvaluator)

Evaluates the fitness of a population using the given evaluator.

GlobalOptimization.evaluate_fitness! — Method

evaluate_fitness!(swarm::Swarm{T}, evaluator::BatchEvaluator{T})

Evaluates the fitness of each candidate in the swarm swarm using the evaluator. Updates the swarms best candidates if any are found.

GlobalOptimization.evaluate_with_penalty — Method

evaluate_with_penalty(evaluator::FeasibilityHandlingEvaluator, candidate::AbstractArray)

GlobalOptimization.feasible — Method

feasible(x, ss::ContinuousRectangularSearchSpace)

Returns true if the point x is feasible in the search space ss, otherwise returns false.

Arguments

x::AbstractVector{T}: the point to check for feasibility.

ss::ContinuousRectangularSearchSpace{T}: the search space to check for feasibility in.

Returns

Bool: true if x is in ss, otherwise false.

Throws

DimensionMismatch: if x does not have the same number of dimensions as ss.

GlobalOptimization.fitness — Method

fitness(c::AbstractCandidate)

Returns the fitness of the candidate c.

GlobalOptimization.fitness — Method

fitness(pop::AbstractPopulation, [i::Integer])

Returns the fitness of the candidates from a population. If i is specified, returns the i-th fitness.

GlobalOptimization.get_best_candidate_in_selection — Method

get_best_candidate_in_selection(population::DEPopulation, idxs)

Get the best candidate in the selected subset of population (as specified by the indices in idxs).

GlobalOptimization.get_display — Method

get_display(opts::AbstractOptions)

Returns the display option from an options type.

GlobalOptimization.get_display_interval — Method

get_display_interval(opts::AbstractAlgorithmSpecificOptions)

Returns the display interval from an algorithm options type.

GlobalOptimization.get_function_value_check — Method

get_function_value_check(opts::AbstractAlgorithmSpecificOptions)

Returns the function value check option from an algorithm options type.

GlobalOptimization.get_general — Method

get_general(opts::AbstractAlgorithmSpecificOptions)

Returns the general options from an algorithm options type.

GlobalOptimization.get_max_time — Method

get_max_time(opts::AbstractAlgorithmSpecificOptions)

Returns the max time option from an algorithm options type.

GlobalOptimization.get_min_cost — Method

get_min_cost(opts::AbstractAlgorithmSpecificOptions)

Returns the min cost option from an algorithm options type.

GlobalOptimization.get_scalar_function — Method

get_scalar_function(prob::AbstractProblem)

Returns cost function plus the infeasibility penalty squared as a scalar value.
This is used for PSO (GA, Differential Evolution, etc. if we ever get around to adding those)

GlobalOptimization.get_scalar_function_with_penalty — Method

get_scalar_function_with_penalty(prob::AbstractProblem)

GlobalOptimization.has_gradient — Method

has_gradient(evaluator::AbstractEvaluator)

Returns true if the evaluator has a gradient, otherwise, false.

GlobalOptimization.initialize! — Method

initialize!(hopper::BasicHopper{T}, search_space::ContinuousRectangularSearchSpace{T})

Initializes the hopper hopper with a uniform distribution in the search space.

GlobalOptimization.initialize! — Method

initialize!(
    swarm::Swarm{T},
    pop_init_method::AbstractPopulationInitialization,
    search_space::ContinuousRectangularSearchSpace{T}
)

Initializes the swarm population with pop_init_method in the search_space.

GlobalOptimization.initialize_fitness! — Method

initialize_fitness!(swarm::Swarm{T}, evaluator::BatchEvaluator{T})

Initializes the fitness of each candidate in the swarm swarm using the evaluator.

GlobalOptimization.intersection — Method

intersection(
    ss1::ContinuousRectangularSearchSpace{T1},
    ss2::ContinuousRectangularSearchSpace{T2}
)

Returns the intersection of the two search spaces ss1 and ss2 as a new search space.

Arguments

ss1::ContinuousRectangularSearchSpace{T1}

ss2::ContinuousRectangularSearchSpace{T2}

Returns

`ContinuousRectangularSearchSpace{promote_type(T1, T2)}

Throws

DimensionMismatch: if ss1 and ss2 do not have the same number of dimensions.

GlobalOptimization.mutate! — Method

mutate!(population::DEPopulation{T}, F)

Mutates the population population using the DE mutation strategy.

This is an implementation of the unified mutation strategy proposed by Ji Qiang and Chad Mitchell in "A Unified Differential Evolution Algorithm for Global Optimization".

GlobalOptimization.num_dims — Method

num_dims(ss::ContinuousRectangularSearchSpace)

Returns the number of dimensions in the search space ss.

Arguments

ss::ContinuousRectangularSearchSpace

Returns

Integer

GlobalOptimization.num_dims — Method

num_dims(prob::AbstractProblem)

Returns the number of dimensions of the decision vector of the optimization problem prob.

GlobalOptimization.push! — Method

push!(step_memory::MBHStepMemoory{T}, step::Vector{T}, pre_step_fitness::T, post_step_fitness::T)

Pushes a step into the step memory.

GlobalOptimization.push_accepted_step! — Method

push_accepted_step!(
    dist::MBHAdaptiveDistribution{T},
    step::AbstractVector{T},
    pre_step_fitness::T,
    post_step_fitness::T,
) where {T}

GlobalOptimization.scalar_function — Method

scalar_function(prob::OptimizationProblem, x::AbstractArray)

Evaluates the objective function f of the optimization problem prob at x and returns the cost function plus half the infeasibility squared.

GlobalOptimization.scalar_function — Method

scalar_function(prob::AbstractNonlinearEquationProblem, x::AbstractArray)

Evaluates the set of nonlinear equations f and returns the nonlinear least squares cost plus half the infeasibility squared.

GlobalOptimization.scalar_function_with_penalty — Method

scalar_function_with_penalty(prob::OptimizationProblem, x::AbstractArray)

Evaluates the objective function f of the optimization problem prob at x and returns the cost function and the infeasibility penalty term as tuple. i.e., for an OptimizationProblem, this is simply the original function.

GlobalOptimization.scalar_function_with_penalty — Method

scalar_function_with_penalty(prob::AbstractNonlinearEquationProblem, x::AbstractArray)

Evaluates the set of nonlinear equations f and returns the nonlinear least squares cost and the infeasibility penalty term as a tuple.

GlobalOptimization.search_space — Method

search_space(prob::AbstractProblem)

Returns the search space of the optimization problem prob.

GlobalOptimization.selection! — Method

selection!(population::DEPopulation{T}, evaluator::BatchEvaluator{T})

Replace candidates with mutants if they have a better fitness.

GlobalOptimization.set_fitness! — Method

set_fitness!(c::AbstractCandidate, fitness)

GlobalOptimization.set_fitness! — Method

set_fitness(pop::AbstractPopulation, fitness, [i::Integer])

Sets the fitness of the candidates from a population. If i is specified, sets the i-th fitness.

GlobalOptimization.stalled — Method

stalled(pso::PSO, stall_value)

Returns true if the PSO algorithm pso is stalled, false otherwise.

GlobalOptimization.step! — Method

step!(swarm::Swarm)

Steps the swarm swarm forward one iteration.

GlobalOptimization.step_std — Method

step_std(step_memory::MBHStepMemory{T}, var_idx::Integer)

Returns the standard deviation of the step memory. If var_idx is specified, then the standard deviation of the step memory for the variable at index var_idx is returned.

GlobalOptimization.update_fitness! — Method

update_fitness!(hopper::BasicHopper{T}, distribution::AbstractMBHDistribution{T})

Updates the hopper fitness information after previously evaluating the fitness of the hopper.

GlobalOptimization.update_global_best! — Method

update_global_best!(pso::PSO)

Updates the global best candidate in the PSO algorithm pso if a better candidate is found.

GlobalOptimization.update_inertia — Method

update_inertia(inertia, range, stall_count)

Returns new inertia weight based on the current 'inertia', the range, and stall_count.

GlobalOptimization.update_velocity! — Method

update_velocity!(swarm::Swarm{T}, cache::Cache, ns::Integer, w, y1, y2)

Updates the velocity of each candidate in the swarm swarm,