Public API Documentation

Documentation for GlobalOptimization's public interface.

Contents

Contents

Index

Optimization Problem

Search Space

Optimization Algorithms

Index

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.SerialPSO

GlobalOptimization.ThreadedPSO

GlobalOptimization.optimize!

Optimization Problem

GlobalOptimization.OptimizationProblem — Type

OptimizationProblem{SS, F}

An optimization problem. Contains the objective function and search space.

Fields

f::F: The objective function.

ss::SS: The search space.

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f, LB, UB)

Constructs an optimization problem with objective function f and a ContinuousRectangularSearchSpace defined by LB and UB.

Arguments

f::F: The objective function.

LB::AbstractVector{<:Real}: The lower bounds of the search space.

UB::AbstractVector{<:Real}: The upper bounds of the search space.

Returns

OptimizationProblem{ContinuousRectangularSearchSpace, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> prob = OptimizationProblem(f, LB, UB)
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f::F, ss::SS)

Constructs an optimization problem with objective function f and search space ss.

Arguments

f::F: The objective function.

ss::SS: The search space.

Returns

OptimizationProblem{SS, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = OptimizationProblem(f, ss)
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

Search Space

GlobalOptimization.ContinuousRectangularSearchSpace — Type

ContinuousRectangularSearchSpace{T <: AbstractFloat}

A RectangularSearchSpace formed by a single continuous set.

Fields

dimmin::Vector{T}: A vector of minimum values for each dimension.

dimmax::Vector{T}: A vector of maximum values for each dimension.

dimdelta::Vector{T}: A vector of the difference between the maximum and minimum values for each dimension.

GlobalOptimization.ContinuousRectangularSearchSpace — Method

ContinuousRectangularSearchSpace(dimmin::AbstractVector{T}, dimmax::AbstractVector{T})

Constructs a new ContinuousRectangularSearchSpace with minimum values dimmin and maximum values dimmax.

Arguments

dimmin::AbstractVector{T}: A vector of minimum values for each dimension.

dimmax::AbstractVector{T}: A vector of maximum values for each dimension.

Returns

ContinuousRectangularSearchSpace{T}

Examples

julia> using GlobalOptimization;
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB)
ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0])

Optimization Algorithms

GlobalOptimization.optimize! — Method

optimize!(opt::AbstractOptimizer)

Perform optimization using the optimizer opt. Returns the results of the optimization.

Arguments

opt::AbstractOptimizer: The optimizer to use.

Returns

Results: The results of the optimization.

GlobalOptimization.SerialPSO — Method

SerialPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a SerialBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractOptimizationProblem{SS}: The optimization problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

GlobalOptimization.ThreadedPSO — Method

ThreadedPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a ThreadedBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractOptimizationProblem{SS}: The optimization problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.