Public API Documentation

Documentation for GlobalOptimization's public interface.

Contents

Contents

Index

Optimization Problem

Search Space

Optimization

Population Initialization

Algorithms

Particle Swarm Optimization

Differential Evolution

Index

GlobalOptimization.Best1

GlobalOptimization.Best2

GlobalOptimization.BinomialCrossoverParameters

GlobalOptimization.BinomialCrossoverParameters

GlobalOptimization.BinomialCrossoverParameters

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.ContinuousRectangularSearchSpace

GlobalOptimization.CovarianceTransformation

GlobalOptimization.CovarianceTransformation

GlobalOptimization.CurrentToBest1

GlobalOptimization.CurrentToBest2

GlobalOptimization.CurrentToRand1

GlobalOptimization.CurrentToRand2

GlobalOptimization.LatinHypercubeInitialization

GlobalOptimization.LatinHypercubeInitialization

GlobalOptimization.MutationParameters

GlobalOptimization.MutationParameters

GlobalOptimization.MutationParameters

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.OptimizationProblem

GlobalOptimization.RadiusLimitedSelector

GlobalOptimization.Rand1

GlobalOptimization.Rand2

GlobalOptimization.RandToBest1

GlobalOptimization.RandToBest2

GlobalOptimization.RandomSubsetSelector

GlobalOptimization.SelfBinomialCrossoverParameters

GlobalOptimization.SelfBinomialCrossoverParameters

GlobalOptimization.SelfMutationParameters

GlobalOptimization.SelfMutationParameters

GlobalOptimization.SimpleSelector

GlobalOptimization.Unified

GlobalOptimization.PolyesterDE

GlobalOptimization.PolyesterPSO

GlobalOptimization.SerialDE

GlobalOptimization.SerialPSO

GlobalOptimization.ThreadedDE

GlobalOptimization.ThreadedPSO

GlobalOptimization.optimize!

Optimization Problem

GlobalOptimization.OptimizationProblem — Type

OptimizationProblem{has_penalty, SS, F, G}

An optimization problem. Contains the objective function and search space.

Fields

f::F: The objective function.

g!::G: The gradient of the objective function.

ss::SS: The search space.

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f, [g], LB, UB)

Constructs an optimization problem with objective function f, optional gradient g, and a ContinuousRectangularSearchSpace defined by LB and UB.

Arguments

f::F: The objective function.

LB::AbstractVector{<:Real}: The lower bounds of the search space.

UB::AbstractVector{<:Real}: The upper bounds of the search space.

Returns

OptimizationProblem{ContinuousRectangularSearchSpace, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> prob = OptimizationProblem(f, LB, UB)
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem{has_penalty}(f::F, [g::G], ss::SS)

Constructs an optimization problem with objective function f, optional gradient g, and search space ss. If has_penalty is specified as true, then the objective function must return a Tuple{T,T} for a given x of type AbstractArray{T}.

Arguments

f::F: The objective function.

g::G: The gradient of the objective function.

ss::SS: The search space.

Returns

OptimizationProblem{SS, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB);
julia> prob = OptimizationProblem(f, ss)
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

GlobalOptimization.OptimizationProblem — Method

OptimizationProblem(f, [g], ss)

Constructs an optimization problem with objective function f, optimal gradient g, and a search space.

Arguments

f::F: The objective function.

g::G: The gradient of the objective function.

ss::SS: The search space.

Returns

OptimizationProblem{ContinuousRectangularSearchSpace, F}

Examples

julia> using GlobalOptimization;
julia> f(x) = sum(x.^2); # Simple sphere function
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> prob = OptimizationProblem(f, ContinuousRectangularSearchSpace(LB, UB))
OptimizationProblem{ContinuousRectangularSearchSpace{Float64}, typeof(f)}(f, ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0]))

Search Space

GlobalOptimization.ContinuousRectangularSearchSpace — Type

ContinuousRectangularSearchSpace{T <: AbstractFloat}

A RectangularSearchSpace formed by a single continuous set.

Fields

dim_min::Vector{T}: A vector of minimum values for each dimension.

dim_max::Vector{T}: A vector of maximum values for each dimension.

dim_delta::Vector{T}: A vector of the difference between the maximum and minimum values for each dimension.

GlobalOptimization.ContinuousRectangularSearchSpace — Method

ContinuousRectangularSearchSpace(dim_min::AbstractVector{T}, dim_max::AbstractVector{T})

Constructs a new ContinuousRectangularSearchSpace with minimum values dim_min and maximum values dim_max.

Arguments

dim_min::AbstractVector{T}: A vector of minimum values for each dimension.

dim_max::AbstractVector{T}: A vector of maximum values for each dimension.

Returns

ContinuousRectangularSearchSpace{T}

Examples

julia> using GlobalOptimization;
julia> LB = [-1.0, 0.0];
julia> UB = [ 1.0, 2.0];
julia> ss = ContinuousRectangularSearchSpace(LB, UB)
ContinuousRectangularSearchSpace{Float64}([-1.0, 0.0], [1.0, 2.0], [2.0, 2.0])

Optimization

GlobalOptimization.optimize! — Method

optimize!(opt::AbstractOptimizer)

Perform optimization using the optimizer opt. Returns the results of the optimization.

Arguments

opt::AbstractOptimizer: The optimizer to use.

Returns

Results: The results of the optimization. See the Results docstring for details on its contents.

Example

julia> using GlobalOptimization
julia> f(x) = sum(x.^2) # Simple sphere function
julia> prob = OptimizationProblem(f, [-1.0, 0.0], [1.0, 2.0])
julia> pso = SerialPSO(prob)
julia> results = optimize!(pso)
Results:
 - Best function value: 6.696180996034206e-20
 - Best candidate: [-2.587698010980842e-10, 0.0]
 - Iterations: 26
 - Time: 0.004351139068603516 seconds
 - Exit flag: 3

Population Initialization

GlobalOptimization.LatinHypercubeInitialization — Type

LatinHypercubeInitialization

Initializes a population using optimal Latin hypercube sampling as implemented in LatinHypercubeSampling.jl.

Fields:

gens::Int: Number of GA generations to use to generate the Latin hypercube samples.

rng::U: Random number generator to use for the Latin hypercube sampling.

pop_size::Int: Size of the GA population used to generate the Latin hypercube samples.

n_tour::Int: Number of tours to use in the GA.

p_tour::Float64: Probability of tour to use in the GA.

inter_sample_weight::Float64: Weight of the inter-sample distance in the GA.

periodic_ae::Bool: Whether to use periodic adaptive evolution in the GA.

ae_power::Float64: Power of the adaptive evolution in the GA.

GlobalOptimization.LatinHypercubeInitialization — Method

LatinHypercubeInitialization(gens::Int = 10; kwargs...)

Initializes a Latin hypercube sampling method with the given parameters.

Arguments

gens::Int: Number of GA generations to use to generate the Latin hypercube samples. Defaults to 10.

Keyword Arguments

rng::U: Random number generator to use for the Latin hypercube sampling. Defaults to Random.GLOBAL_RNG.

pop_size::Int: Size of the GA population used to generate the Latin hypercube samples. Defaults to 100.

n_tour::Int: Number of tours to use in the GA. Defaults to 2.

p_tour::Float64: Probability of tour to use in the GA. Defaults to 0.8.

inter_sample_weight::Float64: Weight of the inter-sample distance in the GA. Defaults to 1.0.

periodic_ae::Bool: Whether to use periodic adaptive evolution in the GA. Defaults to false.

ae_power::Float64: Power of the adaptive evolution in the GA. Defaults to 2.0.

Algorithms

Particle Swarm Optimization

GlobalOptimization.SerialPSO — Function

SerialPSO(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a SerialBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

population_init_method::AbstractPopulationInitialization = UniformInitialization(): The method to use for initializing the population.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

GlobalOptimization.ThreadedPSO — Function

ThreadedPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a ThreadedBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractOptimizationProblem{SS}: The optimization problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

population_init_method::AbstractPopulationInitialization = UniformInitialization(): The method to use for initializing the population.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

GlobalOptimization.PolyesterPSO — Function

PolyesterPSO(prob::AbstractOptimizationProblem{SS}; kwargs...)

Constructs a PSO algorithm with the given options that will employ a PolyesterBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractOptimizationProblem{SS}: The optimization problem to solve.

Keyword Arguments

num_particles::Int = 100: The number of particles to use.

population_init_method::AbstractPopulationInitialization = UniformInitialization(): The method to use for initializing the population.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace} = nothing: The initial bounds to use when initializing particle positions.

max_iterations::Int = 1000: The maximum number of iterations to perform.

function_tolerence::AbstractFloat = 1e-6: The function value tolerence to use for stopping criteria.

max_stall_time::Real = Inf: The maximum amount of time to allow for stall time.

max_stall_iterations::Int = 25: The maximum number of stall iterations to allow.

inertia_range::Tuple{AbstractFloat,AbstractFloat} = (0.1, 1.0): The range of allowable inertia weights.

minimum_neighborhood_fraction::AbstractFloat = 0.25: The minimum neighborhood fraction to use.

self_adjustment_weight::Real = 1.49: The self adjustment weight to use.

social_adjustment_weight::Real = 1.49: The social adjustment weight to use.

display::Bool = false: Whether or not to display the status of the algorithm.

display_interval::Int = 1: The display interval to use.

function_value_check::Bool = true: Whether or not to check for bad function values (Inf or NaN).

max_time::Real = 60.0: The maximum amount of time to allow for optimization.

Returns

PSO: The PSO algorithm.

Differential Evolution

GlobalOptimization.SerialDE — Function

SerialDE(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Construct a serial Differential Evolution (DE) algorithm with the given options that will employ a SerialBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_candidates::Integer=100: The number of candidates in the population.

population_init_method::AbstractPopulationInitialization=UniformInitialization(): The population initialization method.

mutation_params::MP=SelfMutationParameters(Rand1()): The mutation strategy parameters.

crossover_params::CP=BinomialCrossoverParameters(0.6): The crossover strategy parameters.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace}=nothing: The initial bounds for the search space.

max_iterations::Integer=1000: The maximum number of iterations.

max_time::Real=60.0: The maximum time to run the algorithm.

function_tolerance::Real=1e-6: The function tolerance for the stall condition.

max_stall_time::Real=60.0: The maximum stall time for the stall condition.

max_stall_iterations::Integer=100: The maximum number of stall iterations for the stall condition.

min_cost::Real=-Inf: The minimum cost for the algorithm to stop.

function_value_check::Bool=true: Whether to check the function value.

display::Bool=true: Whether to display the algorithm status.

display_interval::Integer=1: The interval at which to display the algorithm status.

GlobalOptimization.ThreadedDE — Function

ThreadedDE(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Construct a serial Differential Evolution (DE) algorithm with the given options that will employ a ThreadedBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_candidates::Integer=100: The number of candidates in the population.

population_init_method::AbstractPopulationInitialization=UniformInitialization(): The population initialization method.

mutation_params::MP=SelfMutationParameters(Rand1()): The mutation strategy parameters.

crossover_params::CP=BinomialCrossoverParameters(0.6): The crossover strategy parameters.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace}=nothing: The initial bounds for the search space.

max_iterations::Integer=1000: The maximum number of iterations.

max_time::Real=60.0: The maximum time to run the algorithm.

function_tolerance::Real=1e-6: The function tolerance for the stall condition.

max_stall_time::Real=60.0: The maximum stall time for the stall condition.

max_stall_iterations::Integer=100: The maximum number of stall iterations for the stall condition.

min_cost::Real=-Inf: The minimum cost for the algorithm to stop.

function_value_check::Bool=true: Whether to check the function value.

display::Bool=true: Whether to display the algorithm status.

display_interval::Integer=1: The interval at which to display the algorithm status.

GlobalOptimization.PolyesterDE — Function

PolyesterDE(prob::AbstractProblem{has_penalty,SS}; kwargs...)

Construct a serial Differential Evolution (DE) algorithm with the given options that will employ a PolyesterBatchEvaluator to evaluate the objective function each iteration.

Arguments

prob::AbstractProblem{has_penalty,SS}: The problem to solve.

Keyword Arguments

num_candidates::Integer=100: The number of candidates in the population.

population_init_method::AbstractPopulationInitialization=UniformInitialization(): The population initialization method.

mutation_params::MP=SelfMutationParameters(Rand1()): The mutation strategy parameters.

crossover_params::CP=BinomialCrossoverParameters(0.6): The crossover strategy parameters.

initial_bounds::Union{Nothing,ContinuousRectangularSearchSpace}=nothing: The initial bounds for the search space.

max_iterations::Integer=1000: The maximum number of iterations.

max_time::Real=60.0: The maximum time to run the algorithm.

function_tolerance::Real=1e-6: The function tolerance for the stall condition.

max_stall_time::Real=60.0: The maximum stall time for the stall condition.

max_stall_iterations::Integer=100: The maximum number of stall iterations for the stall condition.

min_cost::Real=-Inf: The minimum cost for the algorithm to stop.

function_value_check::Bool=true: Whether to check the function value.

display::Bool=true: Whether to display the algorithm status.

display_interval::Integer=1: The interval at which to display the algorithm status.

Mutation Parameters

GlobalOptimization.MutationParameters — Type

MutationParameters{
    AS<:AbstractAdaptationStrategy,
    MS<:AbstractMutationStrategy,
    S<:AbstractSelector,
    D,
}

The parameters for a DE mutation strategy that applies to all current and future candidates in the population.

Fields

F1::Float64: The F₁ weight in the unified mutation strategy.

F2::Float64: The F₂ weight in the unified mutation strategy.

F3::Float64: The F₃ weight in the unified mutation strategy.

F4::Float64: The F₄ weight in the unified mutation strategy.

sel<:AbstractSelector: The selector used to select the candidates considered in mutation.

dist<:Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value.

GlobalOptimization.MutationParameters — Method

MutationParameters(F1, F2, F3, F4; sel=SimpleSelector())

Creates a MutationParameters object with the specified (constant) mutation parameters. These constant mutation parameters are used for all candidates in the population and define a unified mutation strategy as defined in Ji Qiang and Chad Mitchell "A Unified Differential Evolution Algorithm for Global Optimization," 2014, https://www.osti.gov/servlets/purl/1163659

Arguments

F1::Float64: The F₁ weight in the unified mutation strategy.

F2::Float64: The F₂ weight in the unified mutation strategy.

F3::Float64: The F₃ weight in the unified mutation strategy.

F4::Float64: The F₄ weight in the unified mutation strategy.

Keyword Arguments

sel::AbstractSelector: The selector used to select the candidates considered in mutation. Defaults to SimpleSelector().

Returns

MutationParameters{NoAdaptation,Unified,typeof(sel),Nothing}: A mutation parameters object with the specified mutation parameters and selector.

Examples

julia> using GlobalOptimization
julia> params = MutationParameters(0.5, 0.5, 0.5, 0.5)
MutationParameters{GlobalOptimization.NoAdaptation, Unified, SimpleSelector, Nothing}(0.5, 0.5, 0.5, 0.5, SimpleSelector(), nothing)

julia> using GlobalOptimization
julia> params = MutationParameters(0.5, 0.5, 0.5, 0.5; sel=RadiusLimitedSelector(2))
MutationParameters{GlobalOptimization.NoAdaptation, Unified, RadiusLimitedSelector, Nothing}(0.5, 0.5, 0.5, 0.5, RadiusLimitedSelector(2, UInt16[0x6cf0, 0x0c33, 0x0001, 0x0000, 0x0560]), nothing)

GlobalOptimization.MutationParameters — Method

MutationParameters(
    strategy::MS;
    dist=default_mutation_dist,
    sel=SimpleSelector(),
)

Creates a MutationParameters object with the specified mutation strategy with mutation parameter random adaptation. The mutation parameters are adaptively sampled from the provided dist, clamped to the range (0, 1].

Arguments

strategy::MS: The mutation strategy to use. This should be one of the mutation strategies defined in this module (e.g., Rand1, Best2, etc.).

Keyword Arguments

dist::Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters each iteration. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value. Defaults to GlobalOptimization.default_mutation_dist, which is a mixture model comprised of two Cauchy distributions, with probability density given by:

$f_{mix}(x; \mu, \sigma) = 0.5 f(x;\mu_1,\sigma_1) + 0.5 f(x;\mu_2,\sigma_2)$.

where $\mu = \{0.65, 1.0\}$ and $\sigma = \{0.1, 0.1\}$.

sel::AbstractSelector: The selector used to select the candidates considered in mutation. Defaults to SimpleSelector().

Returns

MutationParameters{RandomAdaptation,typeof(strategy),typeof(sel),typeof(dist)}: A mutation parameters object with the specified mutation strategy and selector.

Examples

julia> using GlobalOptimization
julia> params = MutationParameters(Rand1())
MutationParameters{GlobalOptimization.RandomAdaptation, Rand1, SimpleSelector, Distributions.MixtureModel{Distributions.Univariate, Distributions.Continuous, Distributions.Cauchy, Distributions.Categorical{Float64, Vector{Float64}}}}(0.0, 1.0, 0.8450801042502032, 0.0, SimpleSelector(), MixtureModel{Distributions.Cauchy}(K = 2)
components[1] (prior = 0.5000): Distributions.Cauchy{Float64}(μ=0.65, σ=0.1)
components[2] (prior = 0.5000): Distributions.Cauchy{Float64}(μ=1.0, σ=0.1)
)

julia> using GlobalOptimization
julia> using Distributions
julia> params = MutationParameters(Rand1(); dist=Normal(0.5, 0.1))
MutationParameters{GlobalOptimization.RandomAdaptation, Rand1, SimpleSelector, Normal{Float64}}(0.0, 1.0, 0.5061103661726901, 0.0, SimpleSelector(), Normal{Float64}(μ=0.5, σ=0.1))

GlobalOptimization.SelfMutationParameters — Type

SelfMutationParameters{
    AS<:AbstractAdaptationStrategy,
    MS<:AbstractMutationStrategy,
    S<:AbstractSelector,
    D,
}

The parameters for a DE mutation strategy that applies a mutation strategy with unique parameters for each candidate in the population.

Fields

Fs::Vector{SVector{4,Float64}}: The mutation parameters for each candidate in the population. Each element of the vector is an SVector{4} containing the F₁, F₂, F₃, and F₄ weights for the unified mutation strategy.

sel<:AbstractSelector: The selector used to select the candidates considered in mutation.

dist<:Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value.

GlobalOptimization.SelfMutationParameters — Method

SelfMutationParameters(
    strategy::MS;
    dist=default_mutation_dist,
    sel=SimpleSelector(),
)

Creates a SelfMutationParameters object with the specified mutation strategy and mutation parameter random adaptation. The mutation parameters are adaptively sampled from the provided dist, clamped to the range (0, 1].

Arguments

strategy::MS: The mutation strategy to use. This should be one of the mutation strategies defined in this module (e.g., Rand1, Best2, etc.).

Keyword Arguments

dist::Distribution{Univariate,Continuous}: The distribution used to adapt the mutation parameters each iteration. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value. Defaults to GlobalOptimization.default_mutation_dist, which is a mixture model comprised of two Cauchy distributions, with probability density given by:

$f_{mix}(x; \mu, \sigma) = 0.5 f(x;\mu_1,\sigma_1) + 0.5 f(x;\mu_2,\sigma_2)$.

where $\mu = \{0.65, 1.0\}$ and $\sigma = \{0.1, 0.1\}$.

sel::AbstractSelector: The selector used to select the candidates considered in mutation. Defaults to SimpleSelector().

Returns

SelfMutationParameters{RandomAdaptation,typeof(strategy),typeof(sel),typeof(dist)}: A mutation parameters object with the specified mutation strategy and selector.

Examples

julia> using GlobalOptimization
julia> params = SelfMutationParameters(Rand1())
SelfMutationParameters{GlobalOptimization.RandomAdaptation, Rand1, SimpleSelector, MixtureModel{Univariate, Continuous, Cauchy, Categorical{Float64, Vector{Float64}}}}(StaticArraysCore.SVector{4, Float64}[], SimpleSelector(), MixtureModel{Cauchy}(K = 2)
components[1] (prior = 0.5000): Cauchy{Float64}(μ=0.65, σ=0.1)
components[2] (prior = 0.5000): Cauchy{Float64}(μ=1.0, σ=0.1)
)

GlobalOptimization.Rand1 — Type

Rand1

The DE/rand/1 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{r_1} + F\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F$ is a scaling factor, and $r_1$, $r_2$, and $r_3$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.Rand2 — Type

Rand2

The DE/rand/2 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{r_1} + F\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right) + F\left(\mathbf{x}_{r_4} - \mathbf{x}_{r_5}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F$ is a scaling factor, and $r_1$, $r_2$, $r_3$, $r_4$, and $r_5$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.Best1 — Type

Best1

The DE/best/1 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{b} + F\left(\mathbf{x}_{r_1} - \mathbf{x}_{r_2}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F$ is a scaling factor, subscript $b$ denotes the best candidate (in terms of the objective/fitness function), and $r_1$ and $r_2$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.Best2 — Type

Best2

The DE/best/2 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{b} + F\left(\mathbf{x}_{r_1} - \mathbf{x}_{r_2}\right) + F\left(\mathbf{x}_{r_3} - \mathbf{x}_{r_4}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F$ is a scaling factor, subscript $b$ denotes the best candidate, and $r_1$, $r_2$, $r_3$, and $r_4$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.CurrentToBest1 — Type

CurrentToBest1

The DE/current-to-best/1 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{i} + F_{cr}\left(\mathbf{x}_{b} - \mathbf{x}_{i}\right) + F\left(\mathbf{x}_{r_1} - \mathbf{x}_{r_2}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_{cs}$ and $F$ are a scaling factors, subscript $b$ denotes the best candidate, and $r_1$ and $r_2$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.CurrentToBest2 — Type

CurrentToBest2

The DE/current-to-best/2 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{i} + F_{cr}\left(\mathbf{x}_{b} - \mathbf{x}_{i}\right) + F\left(\mathbf{x}_{r_1} - \mathbf{x}_{r_2}\right) + F\left(\mathbf{x}_{r_3} - \mathbf{x}_{r_4}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_{cs}$ and $F$ are a scaling factors, subscript $b$ denotes the best candidate, and $r_1$, $r_2$, $r_3$, and $r_4$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.CurrentToRand1 — Type

CurrentToRand1

The DE/current-to-rand/1 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{i} + F_{cr}\left(\mathbf{x}_{r_1} - \mathbf{x}_{i}\right) + F\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_{cs}$ and $F$ are a scaling factors, and $r_1$, $r_2$, and $r_3$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.CurrentToRand2 — Type

CurrentToRand2

The DE/current-to-rand/2 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{i} + F_{cr}\left(\mathbf{x}_{r_1} - \mathbf{x}_{i}\right) + F\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right) + F\left(\mathbf{x}_{r_4} - \mathbf{x}_{r_5}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_{cs}$ and $F$ are a scaling factors, and $r_1$, $r_2$, $r_3$, $r_4$, and $r_5$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.RandToBest1 — Type

RandToBest1

The DE/rand-to-best/1 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{r_1} + F_{cr}\left(\mathbf{x}_{b} - \mathbf{x}_i\right) + F\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_{cs}$ and $F$ are a scaling factors, subscript $b$ denotes the best candidate, and $r_1$, $r_2$, and $r_3$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.RandToBest2 — Type

RandToBest2

The DE/rand-to-best/2 mutation strategy given by:

$\mathbf{v}_i = \mathbf{x}_{r_1} + F_{cr}\left(\mathbf{x}_{b} - \mathbf{x}_i\right) + F\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right) + F\left(\mathbf{x}_{r_4} - \mathbf{x}_{r_5}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_{cs}$ and $F$ are a scaling factors, subscript $b$ denotes the best candidate, and $r_1$, $r_2$, $r_3$, $r_4$, and $r_5$ are randomly selected integers in the set returned by the selector.

GlobalOptimization.Unified — Type

Unified

The unified DE mutation strategy proposed by Ji Qiang and Chad Mitchell in "A Unified Differential Evolution Algorithm for Global Optimization," 2014, https://www.osti.gov/servlets/purl/1163659.

This mutation strategy is given by:

$\mathbf{v}_i = \mathbf{x}_i + F_1\left(\mathbf{x}_b - \mathbf{x}_i\right) + F_2\left(\mathbf{x}_{r_1} - \mathbf{x}_i\right) + F_3\left(\mathbf{x}_{r_2} - \mathbf{x}_{r_3}\right) + F_4\left(\mathbf{x}_{r_4} - \mathbf{x}_{r_5}\right)$

where $\mathbf{v}_i$ is the target ($i$-th) mutant, $\mathbf{x}_j$ denotes the $j$-th candidate, $F_1$, $F_2$, $F_3$, and $F_4$ are scaling factors, subscript $b$ denotes the best candidate, and $r_1$, $r_2$, $r_3$, $r_4$, and $r_5$ are randomly selected integers in the set returned by the selector.

Note that in the underlying implementation, all mutation strategies are implemented with this formulation, where each unique strategy has a different set of $\{F_i : i \in \{1,2,3,4\}\}$ that are set to 0.0.

GlobalOptimization.SimpleSelector — Type

SimpleSelector

A selector that simply selects all candidates in the population.

GlobalOptimization.RadiusLimitedSelector — Type

RadiusLimitedSelector

A selector that selects candidates within a given radius of the target candidate.

For example, for population size of 10 and a radius of 2, the following will be selected for the given target indices:

target = 5 will select [3, 4, 5, 6, 7]

target = 1 will select [9, 10, 1, 2, 3]

target = 9 will select [7, 8, 9, 10, 1]

GlobalOptimization.RandomSubsetSelector — Type

RandomSubsetSelector

A selector that selects a random subset of candidates from the population. The size of the subset is determined by the size parameter.

DE Crossover Strategies

GlobalOptimization.BinomialCrossoverParameters — Type

BinomialCrossoverParameters{
    AS<:AbstractAdaptationStrategy,
    T<:AbstractCrossoverTransformation,
    D,
}

The parameters for a DE binomial crossover strategy.

Fields

CR::Float64: The crossover rate.

transform::T: The transformation to apply to the candidate and mutant.

dist<:Distribution{Univariate,Continuous}: The distribution used to adapt the crossover rate parameter. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value.

GlobalOptimization.BinomialCrossoverParameters — Method

BinomialCrossoverParameters(CR::Float64; transform=NoTransformation())

Creates a BinomialCrossoverParameters object with a fixed crossover rate CR and optional transformation transform.

Arguments

CR::Float64: The crossover rate.

Keyword Arguments

transform::AbstractCrossoverTransformation: The transformation to apply to the candidate and mutant. Defaults to NoTransformation().

Returns

BinomialCrossoverParameters{NoAdaptation,typeof(transform),Nothing}: A BinomialCrossoverParameters object with a fixed crossover rate and the optionally specified transformation.

Examples

julia> using GlobalOptimization
julia> params = BinomialCrossoverParameters(0.5)
BinomialCrossoverParameters{GlobalOptimization.NoAdaptation, GlobalOptimization.NoTransformation, Nothing}(0.5, GlobalOptimization.NoTransformation(), nothing)

julia> using GlobalOptimization
julia> params = BinomialCrossoverParameters(0.5, transform=CovarianceTransformation(0.5, 0.5, 10))
BinomialCrossoverParameters{GlobalOptimization.NoAdaptation, CovarianceTransformation, Nothing}(0.5, CovarianceTransformation(0.5, 0.5, [0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]), nothing)

GlobalOptimization.BinomialCrossoverParameters — Method

BinomialCrossoverParameters(; dist=default_binomial_crossover_dist, transform=NoTransformation())

Creates a BinomialCrossoverParameters object with an adaptive crossover rate and optional transformation transform.

Keyword Arguments

dist::Distribution{Univariate,Continuous}: The distribution used to adapt the crossover rate parameter. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value. Defaults to default_binomial_crossover_dist, which is a mixture model comprised of two Cauchy distributions, with probability density given by:

$f_{mix}(x; \mu, \sigma) = 0.5 f(x;\mu_1,\sigma_1) + 0.5 f(x;\mu_2,\sigma_2)$

where $\mu = \{0.1, 0.95\}$ and $\sigma = \{0.1, 0.1\}$.

transform::AbstractCrossoverTransformation: The transformation to apply to the candidate and mutant. Defaults to NoTransformation().

Returns

BinomialCrossoverParameters{RandomAdaptation,typeof(transform),typeof(dist)}: A BinomialCrossoverParameters object with an adaptive crossover rate and the optionally specified transformation.

Examples

julia> using GlobalOptimization
julia> params = BinomialCrossoverParameters()
BinomialCrossoverParameters{GlobalOptimization.RandomAdaptation, GlobalOptimization.NoTransformation, Distributions.MixtureModel{Distributions.Univariate, Distributions.Continuous, Distributions.Cauchy, Distributions.Categorical{Float64, Vector{Float64}}}}(0.0, GlobalOptimization.NoTransformation(), MixtureModel{Distributions.Cauchy}(K = 2)
components[1] (prior = 0.5000): Distributions.Cauchy{Float64}(μ=0.1, σ=0.1)
components[2] (prior = 0.5000): Distributions.Cauchy{Float64}(μ=0.95, σ=0.1)
)

julia> using GlobalOptimization
julia> params = BinomialCrossoverParameters(transform=CovarianceTransformation(0.5, 0.5, 10))
BinomialCrossoverParameters{GlobalOptimization.RandomAdaptation, CovarianceTransformation, Distributions.MixtureModel{Distributions.Univariate, Distributions.Continuous, Distributions.Cauchy, Distributions.Categorical{Float64, Vector{Float64}}}}(0.0, CovarianceTransformation(0.5, 0.5, [2.195780764e-314 2.2117846174e-314 … 2.1293782266e-314 1.5617889024864e-311; 2.366805627e-314 2.316670011e-314 … 2.3355803934e-314 1.4259811738567e-311; … ; 2.195781025e-314 2.195781096e-314 … 1.4531427176862e-310 1.27319747493e-313; 2.366805627e-314 2.366805627e-314 … 1.0270459628367e-310 2.121995795e-314], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]), MixtureModel{Distributions.Cauchy}(K = 2)
components[1] (prior = 0.5000): Distributions.Cauchy{Float64}(μ=0.1, σ=0.1)
components[2] (prior = 0.5000): Distributions.Cauchy{Float64}(μ=0.95, σ=0.1)
)

julia> using GlobalOptimization
julia> using Distributions
julia> params = BinomialCrossoverParameters(dist=Uniform(0.0, 1.0))
BinomialCrossoverParameters{GlobalOptimization.RandomAdaptation, GlobalOptimization.NoTransformation, Uniform{Float64}}(0.0, GlobalOptimization.NoTransformation(), Uniform{Float64}(a=0.0, b=1.0))

GlobalOptimization.SelfBinomialCrossoverParameters — Type

SelfBinomialCrossoverParameters{
    AS<:AbstractAdaptationStrategy,
    T<:AbstractCrossoverTransformation,
    D,
}

The parameters for a DE self-adaptive binomial crossover strategy.

Fields

CRs::Vector{Float64}: The crossover rates for each candidate in the population.

transform::T: The transformation to apply to the candidate and mutant.

dist<:Distribution{Univariate,Continuous}: The distribution used to adapt the crossover rate parameter. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value.

GlobalOptimization.SelfBinomialCrossoverParameters — Method

SelfBinomialCrossoverParameters(;
    dist=default_binomial_crossover_dist,
    transform=NoTransformation()
)

Creates a SelfBinomialCrossoverParameters object with an adaptive crossover rate for each candidate in the population and an optional transformation transform.

Keyword Arguments

dist::Distribution{Univariate,Continuous}: The distribution used to adapt the crossover rate parameter. Note that this should generally be a distribution from Distributions.jl, but the only strict requirement is that rand(dist) returns a floating point value. Defaults to default_binomial_crossover_dist, which is a mixture model comprised of two Cauchy distributions, with probability density given by:

$f_{mix}(x; \mu, \sigma) = 0.5 f(x;\mu_1,\sigma_1) + 0.5 f(x;\mu_2,\sigma_2)$

where $\mu = \{0.1, 0.95\}$ and $\sigma = \{0.1, 0.1\}$.

transform::AbstractCrossoverTransformation: The transformation to apply to the candidate and mutant. Defaults to NoTransformation().

Returns

SelfBinomialCrossoverParameters{RandomAdaptation,typeof(transform),typeof(dist)}: A SelfBinomialCrossoverParameters object with an adaptive crossover rate for each candidate and the optionally specified transformation.

Examples

julia> using GlobalOptimization
julia> params = SelfBinomialCrossoverParameters()
SelfBinomialCrossoverParameters{GlobalOptimization.RandomAdaptation, GlobalOptimization.NoTransformation, MixtureModel{Univariate, Continuous, Cauchy, Categorical{Float64, Vector{Float64}}}}(Float64[], GlobalOptimization.NoTransformation(), MixtureModel{Cauchy}(K = 2)
components[1] (prior = 0.5000): Cauchy{Float64}(μ=0.1, σ=0.1)
components[2] (prior = 0.5000): Cauchy{Float64}(μ=0.95, σ=0.1)
)

GlobalOptimization.CovarianceTransformation — Type

CovarianceTransformation{T<:AbstractCrossoverTransformation}

A transformation for performing crossover in the eigen-space of the covariance matrix of the best candidates in the population.

This is an implementation of the method proposed by Wang and Li in "Differential Evolution Based on Covariance Matrix Learning and Bimodal Distribution Parameter Setting, " 2014, DOI: 10.1016/j.asoc.2014.01.038.

Fields

ps::Float64: The proportion of candidates to consider in the covariance matrix. That is, for a population size of N, the covariance matrix is calculated using the clamp(ceil(ps * N), 2, N) best candidates.

pb::Float64: The probability of applying the transformation.

B::Matrix{Float64}: The real part of the eigenvectors of the covariance matrix.

ct::Vector{Float64}: The transformed candidate.

mt::Vector{Float64}: The transformed mutant.

GlobalOptimization.CovarianceTransformation — Method

CovarianceTransformation(ps::Float64, pb::Float64, num_dims::Int)

Creates a CovarianceTransformation object with the specified proportion of candidates to consider in the covariance matrix ps, the probability of applying the transformation pb, and the number of dimensions num_dims.

This is an implementation of the method proposed by Wang and Li in "Differential Evolution Based on Covariance Matrix Learning and Bimodal Distribution Parameter Setting, " 2014, DOI: 10.1016/j.asoc.2014.01.038.

Arguments

ps::Float64: The proportion of candidates to consider in the covariance matrix.

pb::Float64: The probability of applying the transformation.

num_dims::Int: The number of dimensions in the search space.

Returns

CovarianceTransformation: A CovarianceTransformation object with the specified parameters.

Examples

julia> using GlobalOptimization
julia> transformation = CovarianceTransformation(0.5, 0.5, 10)
CovarianceTransformation(0.5, 0.5, [2.3352254645e-314 6.3877104275e-314 … 1.0e-323 5.0e-324; 6.3877051114e-314 6.3877104196e-314 … 6.3877054276e-314 6.387705455e-314; … ; 2.3352254645e-314 2.333217732e-314 … 0.0 6.3877095184e-314; 6.387705143e-314 2.130067282e-314 … 6.387705459e-314 6.387705463e-314], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])